requirements.txt
Django==5.2.6
psycopg[binary]==3.2.10
python-dotenv==1.0.1
django-jazzmin==3.0.1
Pillow==10.4.0
